Move Specflow statement
